feat: move base-page-directive & dispatch store init
